+++ native/bench/ckks.cpp
@@ -1,10 +1,10 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license.
 
-#include "benchmark/benchmark.h"
 #include "seal/seal.h"
 #include "seal/util/rlwe.h"
 #include "bench.h"
+
 using namespace benchmark;
 using namespace sealbench;
 using namespace seal;